pyproject.toml
[tool.poetry]
name = "llm_serv"
version = "1.0"
description = "code"
authors = ["Stefan Dumitrescu <dumitrescu.stefan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "2.10.2"
pytest = "8.3.3"
poetry = "1.8.4"
fastapi = "0.115.0"
uvicorn = "0.32.1"
colorama = "0.4.4"
tenacity = "9.0.0"
PyYAML = "^6.0.1"
rich = "^13.9.4"
pillow = "^11.1.0"
dateparser = "^1.2.0"
boto3 = "1.35.99"
openai = "^1.59.7"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
